# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Queue & Team Management API",
    "description": "API для управления очередями, командами и сессиями (RPC-over-REST included 😎)",
    "version": "1.0.1"
  },
...
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Регистрация нового пользователя",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Успешная регистрация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Вход в систему",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный логин",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Обновление токенов",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Новая пара токенов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/teams": {
      "post": {
        "tags": [
          "Teams"
        ],
        "summary": "Создать команду",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Команда создана (возвращает объект с id)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teams/{id}": {
      "get": {
        "tags": [
          "Teams"
        ],
        "summary": "Посмотреть состав команды",
        "parameters": [
          {
            "$ref": "#/components/parameters/PathId"
          }
        ],
        "responses": {
          "200": {
            "description": "Состав команды",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "PathId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "schemas": {
      "UserRegisterDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "UserLoginDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "RefreshDTO": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
      "TokenResponseDTO": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          }
        }
      },
...
      "CreateTeamDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "TeamDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
    }
  }
}